# openapi.yaml
openapi: 3.0.3
info:
  title: Blog API
  version: 1.0.1
paths:
  /posts:
    get:
      summary: Alle Posts abrufen
      tags: [Posts]
      responses:
        "200":
          description: Liste aller Posts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Post"
        "500":
          description: Server Fehler
    post:
      summary: Post erstellen (Admin)
      tags: [Posts]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PostInput"
      responses:
        "201":
          description: Post erstellt
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Post"
        "400":
          description: Ungültige Eingabe
        "401":
          description: Unauthorized
        "409":
          description: Slug bereits vorhanden

  /posts/{slug}:
    get:
      summary: Einzelnen Post abrufen
      tags: [Posts]
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Post gefunden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Post"
        "404":
          description: Post nicht gefunden

    put:
      summary: Post aktualisieren (Admin)
      tags: [Posts]
      security:
        - bearerAuth: []
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PostInput"
      responses:
        "200":
          description: Post aktualisiert
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Post"
        "400":
          description: Ungültige Eingabe
        "401":
          description: Unauthorized
        "404":
          description: Post nicht gefunden
        "409":
          description: Slug bereits vorhanden

    delete:
      summary: Post löschen (Admin)
      tags: [Posts]
      security:
        - bearerAuth: []
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Post gelöscht
        "401":
          description: Unauthorized
        "404":
          description: Post nicht gefunden

  /auth/login:
    post:
      summary: Admin Login
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: JWT Token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TokenResponse"
        "401":
          description: Falsche Credentials

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Post:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        slug:
          type: string
        content:
          type: string
          description: Markdown content
        created_at:
          type: string
          format: date-time
    PostInput:
      type: object
      required: [title, content]
      properties:
        title:
          type: string
        content:
          type: string
    LoginInput:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
        password:
          type: string
    TokenResponse:
      type: object
      properties:
        token:
          type: string
    ErrorResponse:
      type: object
      properties:
        message:
          type: string